[Fix] 팔로워 수, 팔로잉 수, 리뷰 수를 조회하는 기능 - 기존 user 에서 follow 쪽으로 이동
ref : #62
